Revise find instances analysis
Add an Instance class to wrap variable declaration and the top level
element it was declared in.

Find instances in a hierarchical manner instead of just listing them.
For example, if A is a class with an attribute B and there are two
instances of A, then there should be two instances of B, each with one
of the two A instances as its parent.

Finally, have the find instances analysis start from a single top level
element. In printTopLevelElements, assume this is the (first) program
declaration; this is suitable for our test code.