Add effective subtypes analysis (draft)
Add a draft for an effective subtype analysis. Analyses assignments and
populates a set of effective subtypes a reference or interface reference
may have.

Will be later expanded so invocations are also used to infer the set of
effective subtypes. Right now this is TO-DO.

Adapt AST information so analysis can be conducted more easily.

Also add operational example and test code.