STOO: sort instances previous to transformations
Since classes do not have multiple inheritance, we can ensure all the
instances of any given class (including instances of subclasses
extending the class) lie in a single range. This makes it possible to
maintain the instances in global arrays without having to implement
multiple ranges in the array embedding transformation of ST0.